// apps/api/src/reviews/application/use-cases/create-review.use-case.ts
import { Inject, Injectable } from '@nestjs/common';
import { IReviewRepository } from '../../domain/repositories/review.repository';
import { Review } from '../../domain/entities/review.entity';

export interface CreateReviewCommand {
  customerId: string;
  productId: string;
  rating: number;
  comment?: string;
}

@Injectable()
export class CreateReviewUseCase {
  constructor(
    @Inject('IReviewRepository')
    private readonly reviewRepository: IReviewRepository,
  ) {}

  async execute(command: CreateReviewCommand): Promise<Review> {
    const review = new Review(
      '', // ID is generated by DB
      command.customerId,
      command.productId,
      command.rating,
      command.comment || null,
      false, // isVerified is false by default
      new Date(),
    );

    return this.reviewRepository.create(review);
  }
}